+++ cluster/Cargo.toml
@@ -21,3 +21,4 @@ tokio = { version = "0.2", features = ["full"] }
 arrow_deps = { path = "../arrow_deps" }
 futures = "0.3.7"
 bytes = "0.5"
+chrono = "0.4"
