+++ data_types/src/lib.rs
@@ -4,7 +4,6 @@
 
 #![deny(rust_2018_idioms)]
 #![warn(
-    missing_copy_implementations,
     missing_debug_implementations,
     clippy::explicit_iter_loop,
     clippy::use_self
